requirements.txt
pandas==1.3.3
openpyxl==3.0.9
pyodbc==4.0.30
numpy==1.21.2
PyQt5==5.15.6
python-dateutil==2.8.2